Moved timline to its own folder. changes in imports
